separated combineFilters tests
